CHERI: Separate ptr type and value sizes, relate layouts to address space
This commit modifies DataLayout in order to allow specification of
multiple different address spaces for different purposes. Each address
space can specify a different pointer layout. This pointer layout can
also consist of a different pointer type size (IE how much space the
pointer itself takes up in memory) and value size (IE the bits available
for the value of the address).

To obtain this information from the datalayout then we must
differentiate whether we want the type size or the value size. We should
also pass in the address space in which the pointer will exist.