CHERI: Always construct pointers in codegen with an address space
In CHERI, we don't always use address space 0 for pointers. We need to
construct pointers in the address space specified in the data layout
which is specific to the use case of the pointer. IE pointers to globals
will be in the globals address space, pointers to the stack will be in
the alloca address space, and pointers to code will be in the
instruction address space. This change removes the
`CodeGenCx::type_{i8p,ptr_to}` functions which always construct pointers
in address space 0 and instead changes all the call sites so that they
specify an address space. This is work in progress as many of these call
sites simply use the default address space rather than determining
exactly which address space the pointer type should refer to. We also
replace uses of AddressSpace::DATA with an appropriate address space
from the data layout where possible.