CHERI: Split layout size into ty_size and val_size, use for conversion checks
This patch splits the size field of a type layout into both ty_size and
val_size, where ty_size is the size required to represent the type and
val_size is the size of the value that the type can represent. This
means we can use val_size to determine if the value of one type can be
represented with the value of another - IE for converting pointers
to/from usize.

We use this change to modify SizeSkeleton to match val_size when we can,
which fixes an issue on CHERI whereby mem::transmute between usize and a
pointer would fail due to a size mismatch. Instead of failing to compile
what should actually happen is that we generate an invalid
(non-dereferencable) pointer. To do this we also modify the code emitted
for transmuting a non-pointer to a pointer, such that we explicitly set
the address of a null pointer to the value of the non-pointer.