CHERI: Fix emission of ELF flags for RISC-V
Note: this patch is a combination of an upstream fix and a
CHERI-specific change. When rebasing, make sure the CHERI changes are
applied on top of the upstream fix.

This patch ensures that we emit the correct ELF flags for a CHERI RISC-V
target in an object produced by the Rust compiler rather than LLVM, so
we can link without incompatibility errors.