CHERI: Add cfg item for target_pointer_type_width
This patch simply adds an alternative cfg item similar to
target_pointer_width, which reports the full type width of pointers.
This is especially useful when the target specifies a type width that is
greater than the width of the actual pointer address, such as for
CHERI. The std library needs this information as opposed to the address
size to determine whether some information should be packed if we have
the right pointer type size.