CHERI: Account for fixed sigaction layout on Morello
In the libc crate, the definition of the sigaction type was corrected
for Morello, since it previously incorrectly assumed that a union of
function pointers could be safely represented just as a usize. This
fix requires a change in the users of the type in Rust since the API
had to be changed.