CHERI: Use uintptr_t in interfaces in C where required over usize
This commit follows the principle that uintptr_t can be used to
represent the entirety of a pointer type, whereas usize (size_t in C)
really only represents the address part of a pointer. In CHERI,
uintptr_t is bigger than usize, so if nothing else, we must follow the C
interfaces when deciding to use uintptr_t to ensure ABI compatibility.
This change also modifies Rust code that uses uintptr_t, so that we can
specify uintptr_t as a pointer and still have our code compile and work
correctly.